fixed imports, improved code
